refactor: move property component and repath
