+++ src/components/lesseeHome-page/PropertyCard.tsx
